Move file reading functions into common
