+++ tests/tt_metal/tt_metal/deployment/sources.cmake
@@ -2,6 +2,7 @@
 # Module owners should update this file when adding/removing/renaming source files
 
 set(UNIT_TESTS_DEPLOYMENT_SRC
+    deployment_common.cpp
     dram/dram_base.cpp
     dram/test_dram.cpp
     eth/test_eth_bandwidth.cpp
